Moved game logic to own class
